+++ src/main/kotlin/org/example/root_be/domain/applications/domain/VolunteerApplication.kt
@@ -3,7 +3,7 @@ package org.example.root_be.domain.applications.domain
 import jakarta.persistence.*
 import org.example.root_be.domain.role.domain.Role
 import org.example.root_be.domain.user.domain.User
-import org.example.root_be.domain.volunteer.domain.VolunteerPost
+import org.example.root_be.domain.post.domain.VolunteerPost
 import java.time.LocalDateTime
 
 @Entity
