+++ src/main/kotlin/org/example/root_be/domain/post/domain/repository/VolunteerPostRepository.kt
@@ -1,6 +1,6 @@
-package org.example.root_be.domain.volunteer.domain.repository
+package org.example.root_be.domain.post.domain.repository
 
-import org.example.root_be.domain.volunteer.domain.VolunteerPost
+import org.example.root_be.domain.post.domain.VolunteerPost
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface VolunteerPostRepository: JpaRepository<VolunteerPost, Long> {
