+++ src/main/kotlin/org/example/root_be/domain/role/domain/Role.kt
@@ -1,7 +1,7 @@
 package org.example.root_be.domain.role.domain
 
 import jakarta.persistence.*
-import org.example.root_be.domain.volunteer.domain.VolunteerPost
+import org.example.root_be.domain.post.domain.VolunteerPost
 
 @Entity
 @Table(name = "ROLE")
